Start on refactor of our locust test suite.
