+++ performance/prep.py
