+++ requirements.in
@@ -61,7 +61,6 @@ whitenoise==4.1.4
 django-model-utils==3.2.0
 django-redis
 django-prometheus
-eventlet
 future
 sentry-sdk
 raven
