+++ requirements.txt
@@ -44,10 +44,8 @@ django-webpack-loader==0.7.0  # via -r requirements.in
 django==1.11.29           # via -r requirements.in, django-bulk-update, django-hashedfilenamestorage, django-js-reverse, django-model-utils, django-mptt, django-redis, django-s3-storage, djangorestframework-bulk, jsonfield
 djangorestframework-bulk==0.2.1  # via -r requirements.in
 djangorestframework==3.9.1  # via -r requirements.in, djangorestframework-bulk
-dnspython==1.16.0         # via eventlet
 docutils==0.15.2          # via botocore, sphinx
 ebooklib==0.17.1          # via pressurecooker
-eventlet==0.25.1          # via -r requirements.in
 ffmpy==0.2.2              # via pressurecooker
 flower==0.9.3             # via -r requirements.in
 future==0.18.2            # via -r requirements.in
@@ -62,7 +60,6 @@ google-cloud-logging==1.14.0  # via google-cloud-error-reporting
 google-cloud-storage==1.26.0  # via -r requirements.in
 google-resumable-media==0.5.0  # via google-cloud-storage
 googleapis-common-protos[grpc]==1.51.0  # via google-api-core, grpc-google-iam-v1
-greenlet==0.4.15          # via eventlet
 grpc-google-iam-v1==0.11.4  # via google-cloud-kms
 grpcio==1.27.2            # via -r requirements.in, google-api-core, googleapis-common-protos, grpc-google-iam-v1
 gspread==3.0.1            # via -r requirements.in
@@ -86,7 +83,6 @@ markupsafe==1.1.1         # via -r requirements.in, jinja2
 matplotlib==2.2.3         # via pressurecooker
 metaphone==0.6            # via -r requirements.in
 minio==3.0.3              # via -r requirements.in
-monotonic==1.5            # via eventlet
 newrelic==5.6.0.135       # via -r requirements.in
 numpy==1.15.4             # via matplotlib, pressurecooker, wordcloud
 oauth2client==4.1.3       # via -r requirements.in
@@ -125,7 +121,7 @@ s3transfer==0.3.3         # via boto3
 selenium==3.141.0         # via -r requirements.in
 sentry-sdk==0.14.1        # via -r requirements.in
 singledispatch==3.4.0.3   # via -r requirements.in
-six==1.14.0               # via cycler, django-mailgun, ebooklib, eventlet, google-api-core, google-api-python-client, google-auth, google-resumable-media, grpcio, html5lib, le-pycaption, livereload, matplotlib, oauth2client, progressbar2, protobuf, python-dateutil, python-utils, singledispatch, sphinx, validators, xhtml2pdf
+six==1.14.0               # via cycler, django-mailgun, ebooklib, google-api-core, google-api-python-client, google-auth, google-resumable-media, grpcio, html5lib, le-pycaption, livereload, matplotlib, oauth2client, progressbar2, protobuf, python-dateutil, python-utils, singledispatch, sphinx, validators, xhtml2pdf
 snowballstemmer==2.0.0    # via sphinx
 soupsieve==1.9.5          # via -r requirements.in, beautifulsoup4
 sphinx-autobuild==0.7.1   # via -r requirements.in
